update imports in notebook + cleanup morphometrics functions
